Version 3.0
Finish transitions to Vim 7 - there is probably not a single color unchanged. This version in particular changes that blindingly-bright pink for insert and omni completion to be more in line with the rest of the colorscheme.

Falls back to "default" for older vims (if you accidently start one) so no error messages for non existing features are shown. Perfect starting point for you own vim 7 colorscheme.